Add task timeout extension methods.
When testing async code, it's often useful to wait for something to
happen but being prepared for it not to happen because of a fault in
the code under test (or because you've not written the code required to
pass the test yet). So it can be helpful to wait for a reasonable time
and then give up with an error.

These new WithTimeout extension methods for Task make this easy.