Extended default timeout to 5 seconds.
It seems that running XUnit tests in DNX RC1, things often go slow enough that it frequently takes more than 2 seconds for even fairly simple work triggered asynchronously to complete!

This only seems to be an issue with concurrent execution of tests, but that's the default behaviour.

I'm hoping this is just a perf problem with preview bits. Meanwhile, 5s is tolerable.